test(this commit): This commit continues to enhance the command dashboard functionality ...
# Project State

## Current Focus
This commit continues to enhance the command dashboard functionality by improving code readability and test coverage while maintaining existing features.

## Completed
- [x] Simplified conditional logic in `cyberpunk_dashboard.rs` by replacing if-else with direct boolean assignment
- [x] Improved range checking in `desktop.rs` by using `contains()` method for cleaner code
- [x] Removed unused import in `from_toml.rs`
- [x] Enhanced test assertions in `app.rs` by using more specific checks (e.g., `!cmds.is_empty()` instead of `cmds.len() >= 1`)
- [x] Improved test organization by consistently using `_` prefix for unused variables
- [x] Updated command runner tests in `command.rs` to ignore unused output variables
- [x] Refined test assertions in various test files for better clarity and maintainability
---
category: test
scope: this commit
---

+++ tests/button_test.rs
@@ -190,7 +190,7 @@ fn test_framework_button_empty_label_fallback() {
     let area = make_area(20, 3);
     let plane = btn.render(area);
     assert_eq!(plane.cells[0].char, '[');
-    let end_idx = (1 + "Button".len()) as usize;
+    let end_idx = (1 + "Button".len());
     assert_eq!(plane.cells[end_idx].char, ']');
 }
